Modified to take a parameter: the number of measurements to take
Previous commit message lied - the measurement loop takes a count of the
number of measurements to take, with a 0.5s delay between measurements.
So, this allows you to define the duration of the measurement operation.

I'm using this program merely for demonstration purposes, to test the
accuracy of the measurement, and convince myself the invisible measuring
device is working.