Added dynamic-sensor.py, a new approach
I started from multiple-things.py example code.
I modified the FakeGpioHumidity sensor, to use the hcsr04 code, and
check an *actual* sensor.  This includes a background thread that
continues to check the sensor every 3 seconds, and push updates
back to the gateway.  And this seems to work.
However, there's a bug that causes the console to stream tons of useless
messages.
And, it still has than annoying light bulb, didn't really want that.